requirements.txt
alembic==1.14.0
blinker==1.8.2
CacheControl==0.14.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
lxml_html_clean==0.3.1
Mako==1.3.8
MarkupSafe==3.0.2
msgpack==1.1.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1
pywebcopy==7.0.2
requests==2.32.3
setuptools==75.4.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
WTForms==3.2.1
